Move ModBlocks out of the items package (oops); make datagenAll report status code
